requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
pikepdf>=8.0.0
PySide6>=6.6
